add level to completed games
