+++ services/app/assets/js/widgets/components/Game/CompletedGames.jsx
@@ -2,13 +2,14 @@ import React from 'react';
 import moment from 'moment';
 import ResultIcon from './ResultIcon';
 import UserInfo from '../../containers/UserInfo';
+import GameLevelBadge from '../GameLevelBadge';
 
 const CompletedGames = ({ games }) => (
   <div className="table-responsive">
     <table className="table table-sm table-striped border-gray border-top-0 mb-0">
       <thead>
         <tr>
-          {/* <th className="p-3 border-0">Level</th> */}
+          <th className="p-3 border-0">Level</th>
           <th className="p-3 border-0 text-center" colSpan={2}>
             Players
           </th>
@@ -19,14 +20,9 @@ const CompletedGames = ({ games }) => (
       <tbody>
         {games.map(game => (
           <tr key={game.id}>
-            {/* <td className="p-3 align-middle text-nowrap">
-              <div>
-                <span className={`badge badge-pill badge-${levelToClass[game.level]} mr-1`}>
-                  &nbsp;
-                </span>
-                {game.level}
-              </div>
-            </td> */}
+            <td className="p-3 align-middle text-nowrap">
+              <GameLevelBadge level={game.level} />
+            </td>
             <td className="p-3 align-middle text-nowrap cb-username-td text-truncate">
               <div className="d-flex align-items-center">
                 <ResultIcon gameId={game.id} player1={game.players[0]} player2={game.players[1]} />
