+++ services/app/assets/js/widgets/containers/LobbyWidget.jsx
@@ -21,6 +21,7 @@ import CompletedGames from '../components/Game/CompletedGames';
 import CreateGameDialog from '../components/Game/CreateGameDialog';
 import TopPlayersEver from '../components/TopPlayers/TopPlayersEver';
 import TopPlayersPerPeriod from '../components/TopPlayers/TopPlayersPerPeriod';
+import GameLevelBadge from '../components/GameLevelBadge';
 import levelRatio from '../config/levelRatio';
 
 const Players = ({ players }) => {
@@ -48,17 +49,6 @@ const Players = ({ players }) => {
     </>
   );
 };
-
-const GameLevelBadge = ({ level }) => (
-  <div
-    className="text-center"
-    data-toggle="tooltip"
-    data-placement="right"
-    title={level}
-  >
-    <img alt={level} src={`/assets/images/levels/${level}.svg`} />
-  </div>
-);
 
 const isPlayer = (user, game) => !_.isEmpty(_.find(game.players, { id: user.id }));
